Migrations edited.

+++ Services/Order/Core/TrendShop.Order.Application/Features/CQRS/Commands/OrderDetailCommands/CreateOrderDetailCommand.cs
@@ -12,7 +12,7 @@ namespace TrendShop.Order.Application.Features.CQRS.Commands.OrderDetailCommands
         public string ProductName { get; set; }
         public decimal ProductPrice { get; set; }
         public int ProductAmount { get; set; }
-        public decimal ProductTotalPrice { get; set; } 
-        public decimal OrderingID { get; set; }
+        public decimal ProductTotalPrice { get; set; }
+        public int OrderingID { get; set; }
     }
 }
